+++ python/paddle/distributed/passes/ps_trainer_pass.py
@@ -19,7 +19,7 @@ from _collections import defaultdict
 import paddle
 import paddle.fluid.framework as framework
 from paddle.distributed.passes.pass_base import PassBase, register_pass
-from paddle.fluid.transpiler.collective import SingleProcessMultiThread
+from paddle.distributed.transpiler.collective import SingleProcessMultiThread
 from paddle.framework import core
 from paddle.static import Parameter, Program
