+++ python/paddle/distributed/transpiler/ascend_transpiler.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .. import core
+from paddle.framework import core
+
 from . import collective
 
 OpRole = core.op_proto_and_checker_maker.OpRole
